add CPU RAM and json conversion on CPU_TEMP
