+++ INSTRUMENTS/CPU_TEMP/app.txt
