+++ INSTRUMENTS/CPU_TEMP/output.txt
