--- INSTRUMENTS/CPU_TEMP_RAM/CPU_TEMP_RAM.py
@@ -1,40 +1,0 @@
-from flojoy import flojoy, DataContainer
-import psutil
-import pandas as pd
-
-
-@flojoy
-def CPU_TEMP_RAM(dc, params):
-    """
-    The RASPBERRY_PI Node displays informations about the Raspberry pi which is
-    used to run Flojoy such as CPU temperature and memory available on the device
-
-    Parameters :
-    ------------
-    None
-    """
-    memory_info = psutil.virtual_memory()
-    memory_free = memory_info.available
-    memory_free_go = memory_free / (1024**3)
-
-    temperatures = psutil.sensors_temperatures()
-
-    print(temperatures)
-
-    temp = pd.DataFrame.from_dict(temperatures, orient="index")
-
-    print(temp)
-
-    # cpu_key = ""
-    # for key in temperatures.keys():
-    #    for temp in temperatures[key]:
-    #        if temp.label == "CPU temp":
-    #            cpu_key = key
-    #            break
-
-    # temp = temperatures[cpu_key][0].current
-
-    cpu_data = [temp, memory_free_go]
-    data_description = ["CPU Temp (°C)", "RAM Available (Go)"]
-
-    return DataContainer(type="ordered_pair", x=data_description, y=cpu_data)
